requirements.txt
aioredis==1.3.1
argon2-cffi==19.2.0
asgiref==3.2.10
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.6.2
Automat==20.2.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
bootstrap4==0.1.0
certifi==2019.11.28
cffi==1.13.2
channels==2.4.0
channels-redis==3.0.0
constantly==15.1.0
cryptography==2.9.2
daphne==2.2.5
Django==2.2.5
django-bootstrap==0.2.4
django-bootstrap4==1.1.1
django-braces==1.14.0
Faker==3.0.0
hiredis==1.0.1
hyperlink==19.0.0
idna==2.10
incremental==17.5.0
misaka==2.1.1
msgpack==1.0.0
msgpack-python==0.5.6
Pillow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2019.1
redis==2.10.6
service-identity==18.1.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
text-unidecode==1.3
Twisted==20.3.0
txaio==20.4.1
wincertstore==0.2
zope.interface==5.1.0
